Remove dependency on log4php
There is no need to embed a new logger library as we already
have monolog (and our internal logger is still here).

In the meantime, the ability to record logs in the DB is removed
because it was never used in 11 years.

Other changes were needed to keep various SA tools happy about
my changes.

Part of request #21864 Import forgeupgrade into Tuleap

Change-Id: I9742f4c5d12f0d9b9e5aef661cec031d85cff064