[feature] Create dry_run option for POST projects/
This is part of story #21386 dry run REST project creation

When adding the ?dry_run=true to the REST route POST /projects, then the
project creation is not done but it's checked. It returns a 204 if all
is OK, and a 400 if there are errors.

No new check is added, this patch centralize the permissions checks done
for project creation.

Technical point
---------------

ProjectFieldsResource does not use yet the new object because it will
introduce a change in the logic. The update will be done in a dedicated
commit.

Integration test ProjectCreationTest has been removed because it tested
the old SOAP entrypoint. In addition, creation of project is already tested
in REST testsuite and when doing XML imports for REST / SOAP tests.

Change-Id: If102d81eab0938b2df3962b4df6554cd39799b52