[tech] Refactoring errors collectors in multiple objects
This is part of story #21386 dry run REST project creation

XML and REST errors are almost identical but have a few differences (for
instances with trove cats or custom project fields). They have to be
handled differently.

To do this, the checker has been splitted into multiple objects and
collections to ease the customization of the project registration checks.

In this patch, only a technical refactoring was done. No functional
change is expected.

Change-Id: Id9feafc79d00051a73ad8ba78a49125012a97a58