fix: iterations height depends on visible rows
This is a follow-up to git #tuleap/stable/da779ca48845a481615fb01fbeaad702e418a7b3
where the concept of visible/hidden rows has been introduced, but
forgotten in the computation of iteration height. This produce an
unwanted vertical scroll and screw up the display of the widget.

Part of request #22193: Slow performance when there is a lot of closed item to hide

Change-Id: I9d2ba38e4672b62eddbda6299ed263e217050fa0