request #22209: Introduce a basic HEALTHCHECK directive in production Docker images
This allows the container to do some basic checks on itself to determine
if is healthy or not (this let Swarm reschedule the replica if it
appears to be unhealthy).

The start period has been set to generous value to avoid issues on
environment with an heavy load.

To test you can rebuild the Tuleap Community Docker image locally and
use the provided docker-compose.yml file (do nto forget to mount your
sources inside the container).

Change-Id: I359d9f69f8a257043c38e25977811aa0e1e5732d