ProjectIsNotAProgramException needs to be more generic
This is part of story #21792 define Teams associated to Programs

This error is thrown when creating plan, and getting program_increments of
Program.
In this last context, we don't need to talk about "plan", so the error
is now more generic to be used in theses 2 contexts.

How to test:
 - Use GET project/:id/program_increments with an id that's not a
   Program id
   => The error is fairer in this context
 - Use PUT project/:id/program_plan with an id that's not a Program id
   => The error is always understandable in this context

Change-Id: Icebb5c948eba77b9851d6f8a02e3d0dc7c9f712a