Cannot commit in SVN plugin repositories when there is no SVN core repositories
Issue have been introduced in 69bb9c4d1510bc743193f9bad6441a032695c68e.
Call mades to strpos are stricter in PHP 8.0, due to that when the there
is no SVN core root folder on the filesystem the pre commit hooks
consider that every repositories are a "core" one even when that is not
the case.

Distributed Tuleap end to end tests succeed again.

Part of request #22186: Run Tuleap with PHP 8.0

Change-Id: Id05888dd6eae77e86a7abef6bf01ad5228f966f2