request #22218 Make the text format avaible for text edition
Users may be accustomed to using the text format. Markdown should stay as default but text will need to be available for his users too.

Moreover, the canned responses and default values are saved as text format. Those content can be broken when it's put in a markdown text field.

Change-Id: I9cad4e343973bc9ebdf778c962ed6685bf5fc66e