request #22216 : In TTM, clicking on a change status buttons disables shortcuts
In TestManagement, when I press enter on a mark as status button, I can
no longer use shortcuts in Firefox. I need to press enter or click
somewhere else in order to fully recover keyboard navigation.

This is due to a long lasting bug in Firefox, which does not blur buttons when they are dynamically disabled:
https://bugzilla.mozilla.org/show_bug.cgi?id=706773

To prevent this bug, we "manually" blur the pressed button. This is the
recommended behaviour : https://html.spec.whatwg.org/multipage/interaction.html#focus-processing-model

This makes sure shortcuts and keydowns are still available.
In order to keep navigating with the keyboard after pressing a button, the user only need to keep tabbing.

This patch also enhances a bit the visibility of the focused test in
tests list and focused "mark as" buttons.

Change-Id: Ibda221eeed85b19db9ec0c5e514679b3fcaa74cd