A warning is always displayed in the site administration about non executed forgeupgrade migrations
The output of the ForgeUpgrade CLI has changed a bit so it does not pass
the old checks.

Instead of relying on the presence of strings in the output, we now use
the exit code of the command to signal if updates needs to be executed
or not.

Part of request #21864: Import forgeupgrade into Tuleap

Change-Id: Ib26aa0d09a5de74be7e90de00edd3b274d0b7b6d