[bug] Project import with CLI fails with mandatory categories
This is part of story #21386: dry run REST project creation

The permission check is now done in the collection at project creation,
regadring the type of project creation (REST or XML import).

This check was redone in the category updater object, and so introduce
the check for the XML import that must not be done.

Change-Id: I63985e4c729840d53af8f337f360d7661648c658