Deprecate usage of BucketDb::alterTable and allow existing calls to it to succeed
All the existing calls ::alterTable are incorrect:
* they assume th DB il always named "tuleap"
* the verification of the existence of the property before doing the
  alteration generates an invalid SQL query (it looks like "$property"
  is expected to be a predicate)

The issue is visible since ab23715a348c92bd1452b9cd7e7838b67d69ef33 as
the invalid queries now throw exceptions.

Since it looks like the proposed API is hard to use, we deprecate it and
allow existing call to it to succeed.

Change-Id: Ic9aa0be3b09ee3f9fb11b02bfe5c22d51a390c9d