Have a dedicated object to represent an incomplete program
Part of story #21793 configure Program Management

No functional change expected:
- Non project administrator users cannot access Program Management admin
- Access to Program Management admin is denied for projects that are
Teams

Notes to reviewers:
Using dedicated objects is preferable to primitive types (see Primitive
obsession smell [0]) as we can enforce some verifications when building
the object. Here, we enforce verifying that a potential program being
configured (a ProgramInConfigurationIdentifier) cannot be a Team and the
current user must be administrator of it. Using an int, we cannot
enforce this verification everytime. With a dedicated object, the only
way to get this object is to do the verification, so we can be sure it
is done.

VerifyProjectPermission interface may seem useless: we can just call a
method on PFUser. But this method is badly designed: it creates a new
DAO (without injection) and accesses the database, so using it forces us
to use mocks of PFUser. Using an interface lets us have mock-free tests
and isolates the database access from the domain. It is better from an
hexagonal architecture point of view.

There is some duplication in the code with ProgramForManagement for
example; it should be dealt with in future commits.

[0] http://www.jamesshore.com/v2/blog/2005/primitiveobsession

Change-Id: Ifb1519f00de30b5094314de6f56b804880afc7d0