Limit concurrency when using @tuleap/tlp.fetch::recursiveGet
recursiveGet tries to retrieve all the elements in parallel.
When a lot of requests need to be sent, it becomes counterproductive:
* server and/or client might suffer from the quantity of requests to
  process
* browsers might decide to throttle the misbehaving script anyway

With this contribution, recursiveGet is throttle by default and
developers can manage the number of requests they want to sent in
parallel if they have the need for it.

No functional changes are expected.

Part of story #22213: get a document with basic fields

Change-Id: I3d7aa23026ac6fb5c73f577f968186965a16a540