Use ProgramForAdministrationIdentifier instead of int
Part of story #21793 configure Program Management

No functional change expected in Program Management admin page

Notes to reviewers:
Using int has weaker guarantees because we could pass any number. Using
the dedicated object, we are certain that the given project is not a
Team and that current user is administrator of it.

Change-Id: I4f88012a80a97bce1701df51447890c938a6b818