Switch CLI commands and backend tasks to PHP 8.0
All CLI tools and backend/background tasks are now expected to use
PHP 8.0.

A previous attempt was made and then reverted due to issues when
importing Mediawiki contents (detected when importing content for the E2E tests).
Those issues have been resolved and the specfile now enforces that
a version with the fixes is installed.

Part of request #22186: Run Tuleap with PHP 8.0

Change-Id: I8637c8e947526a0f81ebeadbf717915cb6e39dba