Remove generic usage of paths from ForgeUpgrade
When ForgeUpgrade was a generic tool it made sense to have
a way to configure where migration buckets could be found.

Today the paths are constraints, either it's in core or in
one of the installed plugin.

The ability to configure the path to listen is then removed
as well as "ignore pre-up" or "pre up only" that have no
legitimate usage.

Site administrators will be happy has the verification
of "system is up to date" is way more fast (no longer
look in all sub-directories of plugins).

forgeupgrade tools no longer need --config option and
`/etc/tuleap/forgeupgrade/config.ini` is obsolete too
and can be removed.

Part of request #21864: Import forgeupgrade into Tuleap

Change-Id: I119dd18010b6760aa8db96839ee9e8ab69b83fc6