feat: Custom primary action in project created modal
The success modal displayed after project creation can have a custom
primary action. In the context of program management the primary action
should be the ability to configure the teams.

In order to test it you should add the following parameter:
xml_template_name=program_management_program in the url after a project
creation.

It will be used for real later when the UI to create project based on
program will be merged.

Part of story #22210: select SAFe templates at project creation

Change-Id: Iccc952bb21863c02e4ebbfc9f88af62933badd64