fix: use unicode for registered symbol
Instead of the ascii variant (R), we should use ® instead.

Due to the fact that our current gettext extraction does not handle
non-ascii characters in source strings, we have to do a str_replace
afterwards.

This is applied only on SAFe® and Scaled Agile Framework® trademarks.

By the way, since it is a trademark, the correpsonding category should
not be in uppercase, we should keep the lowercased e in SAFe®

Part of story #22210: select SAFe templates at project creation

Change-Id: Id9c3501c613816c6710329ba7fd73a871082151c