Cannot create project based on the "Agile ALM" template
\array_search() can return 0 (which is evaluated as false) so to test
for the existance (or absence) of a key you either need to explicitly
check for \false or use the more appropriate \in_array().

In the present case the issue is deeper than that, the verification is
done twice in two different ways.

This contribution fixes the initial issue by removing the broken
verification and improves the error message so it's easier to understand
for the users what's the actual issue.

Issue introduced by 646be67d7d9fc1a89f1ed3c8fc38f783599337cd.

Part of story #22210: select SAFe templates at project creation

Change-Id: I3bac40bdc85820f8542675a8930793c2ad9c6e5f