Update nixpkgs pin
The version of Cosign available in our current pin of nixpkgs is old and
the signature format has changed. In order to get the latest version of
Cosign [0] for upcoming contributions we update the pin.

Nothing stands out in the upgrades, this is mainly minor version bumps:
* PHP: 7.4.18 -> 7.4.21
* Composer: 2.0.13 -> 2.1.3
* NodeJS: 16.2.0 -> 16.5.0
* Docker Compose: 1.28.6 -> 1.29.2

Part of request #22240: Sign Docker images using Cosign

[0] https://github.com/NixOS/nixpkgs/pull/130070

Change-Id: I9743ca76a9828f1f2a5363dbb5b701a0196e3b57