request #22560 : Error when pressing `escape` in Taskboard
Pressing `escape` in empty space of Taskboard should not trigger an error
When clicking on a "white space" of Taskboard and then pressing `escape`, an error "Active element should have a SWIMLANE parent" appears in the console.

This should not happen as not all element in Taskboard are children of
a swimlane.
If focus in not in taskboard, nothing should happen and default
behaviour of `escape` should not be prevented.

Change-Id: Ie2f66145f780d3daae3940485e1be152051ba50e