feat: Monospace switch in ForumML threads
Copy the behavior that already exists on FlamingParrot version of
threads. User can switch individual messages to a monospaced font.

This continues to add BurningParrot for ML thread introduced at
git #tuleap/stable/d1300af67323b2ebc7f5a3d849ee39d654cc42f1

In order to not produce a too big review, the full conversion to
BurningParrot will be done in multiple commits.

To test, either uncomment the call to OneThreadController::getUrl() in
ThreadsPresenterBuilder so that you can access the new view when
browsing threads, or build the url by yourself:
  /plugins/forumml/list/:list-id/threads/:thread-id

Things to be done in next commits:
* Reply to a message
* Remove commented call and use OneThreadController::getUrl() everywhere
so that we can delete legacy code

Part of request #18846: Switch ForumML to BurningParrot

Change-Id: I9ff5de6632aa06751b84c10e459448f68482da20