feat: export dashboards in XML
This is a first step to be able to export project dashboards.
For now it only exports the dashboards and their names. Widgets will be
done in a dedicated changeset.

RelaxNG definition for dashboards has been updated to be able to export
empty dashboards (current definition forced us to have at least one
widget).

Exported projects with or without dashboards can be imported back.

Part of request #22567: Add dashboards configurations at project export

Change-Id: I88abbaded1f26af1206cad1137e97000de2b3bb1