Inject the function doing the transformation of the document to the DOCX format as a dependency
This makes the separation clear between what's it's unlikely to change
(the retrieval of information and how we want to export it) and the
section that does the transformation to the targetted format.

We might need to switch the underlying implementation in case of issues.

Part of story #22213: get a document with basic fields

Change-Id: I98066e0fda1fc811382b6e351a7c843d3acefe45