Add a TOC to the generated document
It's likely the element does not show content when you open it.
However if you reload the index, the TOC should show up.

The element to be added to the TOC are styled with a dedicated style
because it appears it's the best way to have something working in major
tools (at least, that's the case for LibreOffice).

Part of story #22213: get a document with basic fields

Change-Id: I0070cbce26478f92887bfde2ee5b7f6077599199