Preload most used plugins
The list of selected plugins was infered by the amount of time
autoloader was looking for files in them. The overall gain doesn't seems
as high as expected on my laptop (~7% on an artifact view) but I bet
that it will be way more interesting on server with poor filesystem
performances.

In order to start your server you will need to raise the memory limit to
256M in **php.ini** (AFAIU, preload is done before php-fpm full startup
so fpm configuration is not relevant for preloading).

Part of request #22563 [Performances] preload Tuleap files

Change-Id: Id03ebb35a8320d80cde286d108a2970eb9217021