XML import optional labels customisations
part of: story #22210 select SAFe templates at project creation

Note:
We do not want to provide custom names by defautl for our Program
template. Hence, there is no `<customisation/>` tag in the provided
configuration. The goal of this commit is to handle the customisation of
the PI section and milestones names if there are defined in the config
file.

How to test:
- Edit `plugins/program_management/resources/templates/program_management_config.xml`
  and add this tag after the `<configuration/>` tag:

``` XML
<customisation>
  <program_increments_section_name>Foo</program_increments_section_name>
  <program_increments_milestones_name>Bar</program_increments_milestones_name>
</customisation>
```

- Run `make generate-templates`
- Create a project based on the Essential SAFe® - ART and Program
  Management template

--> It should be created successfully

- In the creation success modal, click on [Configure teams]
- Add a team to your program to toggle the Program configuration

--> The optional field `PI section custom name` contains `Foo`
--> The optional field `Milestone custom name` contains `Bar`

Change-Id: I50271532d8a17a25e577c6d25453fa933abdc1c4