Prefill the table of contents of the generated document
For now the page number is not added, it will be added later.

Most (all?) of the tools using docx documents do not generate the table
of contents when the document is opened (at least not without an extra
confirmation). This happen for multiple reasons:
* the generation of the TOC can take some non negligible times on large
  documents
* the TOC might refers to external fields so you might not want to
  access them each time you open the document

The strategy adopted by the text editors seems to be to prefill the TOC
with the content. This way, even if the user do not generate it
explicitly, it does not seem broken.

The library used to generate the document does not support this directly
so we re-implement it. However in a effort to (vastly) reduce the complexity
the implementation is specific for our use case.

Part of story #22213: get a document with basic fields

Change-Id: I2e7d35468545507c00fa136f5d85fd378c5b6617