Insert a pageref element for each line of the TOC
This page number is likely to not be visible when the document because
the fields need to be refreshed which is an operation avoided by editors
(especially when the document has been downloaded).

We however still try to add it as it is our best chance to have a
complete TOC without an explicit action of the user.

The PageRef element is not present in the library we use. It will be
contributed back.

Part of story #22213: get a document with basic fields

Change-Id: Ic6a76c3328455f0c7d3222db83d1b69bb620a851