request #22589: Memory limit is reached when uploading a large file via tus
Issue has been introduced by git #tuleap/stable/e94e0aa8b6270f0397b88ac6c9d44be7bf1a19b9
which comes with an upgrade of our PSR7 implementation.
guzzle/psr7:1.8.0+ has introduced a specific handling of the php://input
wrapper to workaround issues on deployment where ext-curl is not
available [0]. In our situation where we can be sure that ext-curl is
available this consumes memory with no benefit.

To avoid that, we explicitly disable this specific handling of the
php://input wrapper by redefining an implementation of a StreamFactory.
Since this is not something that is expected to manipuled directly by
consumers elsewhere in Tuleap, the usage of this implementation is
kept internal to the Tuleap\Http package.

[0] https://github.com/guzzle/psr7/pull/354/commits/0a96a2d9ec9cd589e1fbc482cd37d5451fe51c63

Change-Id: Iec369d4d35e1597bf3f3a6e48654bb79136f77bf