request #22592: Move to npm 7
Since the *shrinkwrap hooks are not executed anymore, an alternative way
to clean up the lockfiles from our local dependencies. The npm CLI is
wrapped inside our own script and we detect that `npm install` or `npm
update` has been called we clean up the lockfile right after.

Also note that we needed to implement another workaround: npm 7
considers dependencies of packages in subfolder has extraneous so it
prunes them automatically (as a result we end up with broken
node_modules). To avoid that we replay a `npm install` in the packages
located in subfolder. npm 7 being a bit faster and the dependencies
being already cached it does not seems to impact the performances in my
local benchmarks.

Since we rewrap the npm CLI, adding an explicit hook in each packages is
no more needed (at least, it will not be forgotten anymore...). The
rewraping of the npm CLI is done automatically for all the users of the
environment provided by the nix environment. Users not relying on nix
will have to find a way to alias `npm` to `tools/utils/nix/build-tools/npm-cli-clean-up-local-tuleap-dep.sh`
when they are working the Tuleap sources.

Change-Id: Ib6ccb58759f3994722f97afccffca7ab34db32f1