request #22596: Drop the C toolchain from the Nix shell dev environment
The dev environment provided by nix-shell is built upon `mkShell`, which
itself comes with a predefined set of dependencies via the standard
build environment [0]. The default stdenv have a complete C toolchain
which we do not need for our day to day work. Dropping it reduces the
size of the closure with no disadvantages for our context.

Before:
```
$ nix path-info --closure-size --human-readable $(nix-build --no-out-link -A inputDerivation shell.nix)
/nix/store/zx4kndaxla7zqflpn8q6w96snfw8v2cc-nix-shell	   1.2G
```

Now:
```
$ nix path-info --closure-size --human-readable $(nix-build --no-out-link -A inputDerivation shell.nix)
/nix/store/m7xsdsknihn9nvqr5prblkwbg3w7hfn7-nix-shell	1021.2M
```

[0] https://nixos.org/manual/nixpkgs/stable/#chap-stdenv

Change-Id: Ieb95390569937de4c0543f27d88b6d358be51945