request #22605: Add a link to the documentation in the systemd units
No functional changes. A link to the documentation [0] is added in the units
description.

https://docs.tuleap.org is used because we cannot determine at this
level where the possibly local documentation installation is. Anyway if
you are looking at the unit the local documentation might not be
accessible.

To test it locally, you need to manually redeploy the units and tell
systemd to reload them with `systemctl daemon-reload`.

[0] https://www.freedesktop.org/software/systemd/man/systemd.unit.html#Documentation=

Change-Id: I076522ec544030e6359c356eeaa712bea6109871