request #22611 Taskboard is broken when stories are linked to tasks that are in another project
How to reproduce:
 - Have a task in project A
 - Have an US in project B
 - Link task to US as a child
 - Go to taskboard in project B

Expected results:
 - The US is displayed like a solo card
 - There is no empty column at the left
 - The ccounter for each column is ok

Info for reviewer:
I chose to solve it by changing the back-end, because taskboard is
not really broken, it runs as expected, but the data sent by the back
are wrong.
So, instead of fixing the front and applying a big fix, I decided to change
the back to return good data.

Change-Id: If159bf708fcf8d306de5292a28f92620ec90ebf1