Service icons are displayed directly using FontAwesome 5
The shim is no more needed for them.

To avoid a potentially dangerous DB migration, the existing icon ID are
kept as is. Instead of directly the icon ID we now search in lookup
table the corresponding FA icon. This will also makes easier future
migrations to new versions because we do not need to strictly adhere to
the FA naming system.

The easiest way to test the patch is to open the add/edit modal of the
service and to check if all the icons are properly displayed in the
selectbox. Note there is an issue with the TripAdvisor icon which was
introduced by 78cb0e1b3dc34eb0a81fab1935b2426eea41fa02 and will dealt
with in a dedicated contribution.

Part of request #17110: Update to Font Awesome 5

Change-Id: Ic815d086217851b994287f40fbb2a758a5c45e5e