Remove javascript:help_window handler
The `javascript:` URIs need to go so we can deploy a strict
Content-Security-Policy.

The existing usages of javascript:help_window are replaced by event
listeners on a[data-help-window] elements. Those event listeners are
only added on FP pages because there are no usages of
`javascript:help_window` on BP pages (and that should stay that way).

Part of request #17967: Deploy a useful content security policy

Change-Id: I21fda34d690d371a0a1717338485bbee95dc186d