request #22615 Transition rule for add to top backlog can fail
Introduced by 277446571faba213518c95f6be1dd457d83209f4

How to test:
- Disallow anonymous user in site administration access.
- In a plannable tracker in a Program project, edit the workflow
transitions. Add an "Add to top backlog" post action in any
transition. Save it and open the same transition again. The post action
should be saved and loaded correctly.
- No functional change expected when planning a feature with mass
change, drag and drop in Program or drag and drop in kanban with the
workflow enabled.
- No functional change expected in the rest of Program Management.

Notes:
- Entry point is in TopBacklogChangeProcessor
- The "Bypass" is a nullable and not a boolean so we can describe why it
exists. For now, we don't need to do anything with it so it's empty.

Change-Id: Ifb4b2ff7bedb5941ff8f423c9625f1f7ba9f098d