Group fields of the artifacts by fieldsets in the generated document
The elements outside of a fieldset are grouped together and are
displayed before the content of the fieldsets.
The fields of each fieldset are ordered according to the order defined
by the tracker structure.

A future contribution will set the label of each fieldset as a subtitle
of the appropriate level. For now they are only displayed as a basic
paragraph text.

Part of story #22568: have a structured document

Change-Id: Ic5f127b44f9b63fc056f0beb59ae5e0223133d6b