Fieldset titles are used as subtitles
The numbering scheme has been hardcoded to have a maximum depth of 10
elements after that it will fallback to the default numbering scheme.
This means that if more than 8 fieldsets are nested together the titles
will not have the same format than the others. It is unlikely to be an
issue since nesting too many fieldsets make the whole thing barely
usable in the web UI.

Part of story #22568: have a structured document

Change-Id: If5139e26e5c1a6efec17d5a678f3ce33eb13e0a8