Tracker reports are not imported correctly
Issue introduced by git #tuleap/stable/56dc87cf1c1f2489b4acfaff70d2b071dc8bbbbd.

When doing a XML import, the reports being created are assigned a string
with a random part as an ID. The field as been set as an integer only
field which cause the string to be casted and result in a not correct
import.

Part of story #22568: have a structured document

Change-Id: I144d9b7600faa10dd5d849bbb86cc0023006b22c