refactor: Convert Issue XML template tracker to PHP code
This commit showcases the usage of PHP code instead of XML file to
define the structure of the tracker inside the "Issue tracking" project
template.

The intent is to increase readability and maintenance of templates.

We make the assumption that "errors" introduced in the past in XML
templates will be more easily spotted via PHP code. For example in the
following code, we can see that there is an "issue" with permissions (no
need to define update permission on project members if registered can
update)

```php
(XMLSelectBoxField::fromTrackerAndName($tracker, 'status'))
    ->withLabel('Status')
    ->withRank(0)
    ->withPermissions(
	new ReadPermission('UGROUP_ANONYMOUS'),
	new UpdatePermission('UGROUP_REGISTERED'),
	new SubmitPermission('UGROUP_REGISTERED'),
	new UpdatePermission('UGROUP_PROJECT_MEMBERS')
    )
```

compared to:

```xml
  …
  <permission scope="field" REF="F13609" ugroup="UGROUP_REGISTERED" type="PLUGIN_TRACKER_FIELD_SUBMIT"/>
  <permission scope="field" REF="F13609" ugroup="UGROUP_REGISTERED" type="PLUGIN_TRACKER_FIELD_UPDATE"/>
  <permission scope="field" REF="F13617" ugroup="UGROUP_ANONYMOUS" type="PLUGIN_TRACKER_FIELD_READ"/>
  <permission scope="field" REF="F13617" ugroup="UGROUP_REGISTERED" type="PLUGIN_TRACKER_FIELD_UPDATE"/>
  <permission scope="field" REF="F13617" ugroup="UGROUP_REGISTERED" type="PLUGIN_TRACKER_FIELD_SUBMIT"/>
  <permission scope="field" REF="F13617" ugroup="UGROUP_PROJECT_MEMBERS" type="PLUGIN_TRACKER_FIELD_UPDATE"/>
  <permission scope="field" REF="F13618" ugroup="UGROUP_ANONYMOUS" type="PLUGIN_TRACKER_FIELD_READ"/>
  <permission scope="field" REF="F13618" ugroup="UGROUP_REGISTERED" type="PLUGIN_TRACKER_FIELD_SUBMIT"/>
  …
```

Note: since the intent of the commit is to introduce PHP code and not to
adjust the template, PHP code might still contains such weird behavior.

Old (XML) tracker definition:
* tools/utils/setup_templates/issues/issue_template.xml

New (PHP) tracker definition, thanks to a dedicated tool to help the
conversion[0]:
* plugins/tracker/include/Tracker/Template/IssuesTemplate.php
* plugins/cardwall/include/XML/Template/ComplementIssuesTemplate.php
* plugins/graphontrackersv5/include/XML/Template/ComplementIssuesTemplate.php

The drawback of the current proposal is that definition of the template
is now split accross multiple files (so that plugins do not leak in each
others). However it will be easier now to find usage of fields via their
constants (as shown in this commit with STATUS_FIELD_ID).

For now only the tracker is converted. Dashboards will be done in a
dedicated contribution.

Suggested entry point of the contribution:
\Tuleap\Project\Registration\Template\IssuesTemplate::getXMLPath()

An "Issue tracking" project created with this contribution should have
nearly the same structure than one created before. The only difference
is the definition of the Workflow. Since it is marked as unused, the
definitions of transitions and states have been omitted here.

Jira import should not suffer any functional changes (we reused and
complemented code to generate XML).

Part of request #22651: Convert Issue XML template to PHP code

[0]: See https://gerrit.tuleap.net/c/tuleap/+/23751

Change-Id: Ie952a4dd98ebd433956632fc1b47c37e097936ff