[bug] Fix header X-PAGINATION-SIZE for user_groups/:id/users
This fixes request #22206 [Bug] response headers the "X-PAGINATION-SIZE" give the number of users with status="A" only for dynamic user groups

In the response headers the "X-PAGINATION-SIZE" give the number of users with status "A" or "R"
but the response content is giving all users with status "A" or "R" or "S"

The issue for is only for dynamic user groups.
Static user groups does not return the suspended users.

Change-Id: I22beb2844edcc2c87007fec0ffb2d5eee172dfbc