Replace FormatSelector by web component
part of request #22647 Convert Vue components to hybrids web components

How to test:
No functional change expected when selecting a format for a text field
or a follow-up comment. When you select a format, the text field will
adapt (for example it'll create a CKEditor in HTML, the buttons will
appear in Markdown).

Notes:
I also added the unit test I had marked as optional suggestion in the
Radio Buttons review.

Change-Id: If24af8872a2f40b962f70becd606f6b0d2afb5ff