request #22680: Manage typechecking and .d.ts generation with Vite plugins
This reduce a bit the amount of boilerplate needed when creating a new
library and centralize the configuration.

The .d.ts files are generated with vite-dts [0]. It does that by
re-exporting the entry module which is nice in our context since the
sources are always available.

The typechecking is handled via vite-plugin-checker [1] which work a bit
like the fork-ts-checker-webpack-plugin in the Webpack world.

[0] https://github.com/alloc/vite-dts
[1] https://github.com/fi3ework/vite-plugin-checker

Change-Id: I4b414b1284dcf9e6bfeea26fa4d1263f4e7cb63e