Propose parent from program
With this contribution, a user story created in a Team should be able to
be attached to a Feature (or a Program Enabler) created in a parent
Program.

The contribution is a bit large because of the introduction of the
interface on PlanningFactory to be able to write tests against
getRootPlanning.

I also had to move part of the logic that was in ArtifactLink class into
the PossibleParent event or Tracker_Artifact_PossibleParentsRetriever.
The main change here is that PossibleParent event no longer deal with a
"Parent Tracker" but with the source tracker itself. In the context of
Program Management there is no cross project hierarchy, hence no Parent
Tracker. However the configuration of the project themselves allows to
propose parents.

Final note: from a pure Process point of view it might be wrong to
attach a User Story to a Program Enabler but:
* This has no impact on Tuleap so it's rather safe from a data
  consitency point of view.
* We don't distinguish the concept of "hierarchy of tracker" anymore so
  we cannot "force" the possible parents.

ATM, we choose to keep the implementation "simple". If there is a strong
need we will treat it later on.

Part of story #22681 select a parent from Program for my features/enablers

Change-Id: I40d4075c7b7f5e8c839b66ef0f7927232bbbd708